# docs/INF/openapi/teralya-openapi-v1.0.yaml
openapi: 3.1.0
...
info:
  title: Teralya API
  version: 1.0.0
  description: Contrato OpenAPI 3.1 del MVP. Derivado de INF-08 v2.5, INF-10 v1.0 e INF-10-A v1.0.
  contact:
    name: Teralya
    url: https://teralya.es
  license:
    name: Propietario — Teralya
    identifier: LicenseRef-Teralya-Proprietary
servers:
- url: https://api.teralya.es
  description: Producción
...
paths:
...
  /admin/incidencias:
    get:
      operationId: api040
      x-teralya-api-code: API-040
      x-teralya-cu:
      - CU-027
      x-teralya-pt:
      - PT-ADM-008
      tags:
      - Administración
      summary: Listar incidencias
      description: Contrato normativo de API-040.
      security:
      - BearerOpaque: []
      responses:
        '200':
          description: Operación completada
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageIncidentSummary'
              examples:
                success:
                  value:
                    items: []
                    page: 1
                    page_size: 1
                    total_items: 0
                    total_pages: 0
        '400':
          description: Validation Error
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/validation-error
                    title: Validation Error
                    status: 400
                    detail: La solicitud no pudo completarse.
                    code: VALIDATION_ERROR
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '401':
          description: Authentication Required
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            WWW-Authenticate:
              description: Desafío Bearer según RFC 6750
              schema:
                type: string
                const: Bearer
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/authentication-required
                    title: Authentication Required
                    status: 401
                    detail: La solicitud no pudo completarse.
                    code: AUTHENTICATION_REQUIRED
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '403':
          description: Forbidden
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/forbidden
                    title: Forbidden
                    status: 403
                    detail: La solicitud no pudo completarse.
                    code: FORBIDDEN
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '500':
          description: Internal Error
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/internal-error
                    title: Internal Error
                    status: 500
                    detail: La solicitud no pudo completarse.
                    code: INTERNAL_ERROR
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
      parameters:
      - name: estado
        in: query
        required: false
        schema: &id032
          type: string
          enum:
          - abierta
          - en_revision
          - resuelta
          - cerrada
        example: abierta
      - *id004
      - *id005
  /admin/incidencias/{id}:
    get:
      operationId: api041
      x-teralya-api-code: API-041
      x-teralya-cu:
      - CU-027
      x-teralya-pt:
      - PT-ADM-009
      tags:
      - Administración
      summary: Consultar incidencia
      description: Contrato normativo de API-041.
      security:
      - BearerOpaque: []
      responses:
        '200':
          description: Operación completada
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentDetail'
              examples:
                success:
                  value:
                    id: 11111111-1111-4111-8111-111111111111
                    tipo: aaaaa
                    estado: abierta
                    fecha: '2026-07-13T12:00:00Z'
                    recurso_relacionado:
                      tipo: pedido
                      id: 11111111-1111-4111-8111-111111111111
                    descripcion: aaaaa
                    updated_at: '2026-07-13T12:00:00Z'
        '401':
          description: Authentication Required
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            WWW-Authenticate:
              description: Desafío Bearer según RFC 6750
              schema:
                type: string
                const: Bearer
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/authentication-required
                    title: Authentication Required
                    status: 401
                    detail: La solicitud no pudo completarse.
                    code: AUTHENTICATION_REQUIRED
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '403':
          description: Forbidden
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/forbidden
                    title: Forbidden
                    status: 403
                    detail: La solicitud no pudo completarse.
                    code: FORBIDDEN
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '404':
          description: Resource Not Found
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/resource-not-found
                    title: Resource Not Found
                    status: 404
                    detail: La solicitud no pudo completarse.
                    code: RESOURCE_NOT_FOUND
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '500':
          description: Internal Error
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/internal-error
                    title: Internal Error
                    status: 500
                    detail: La solicitud no pudo completarse.
                    code: INTERNAL_ERROR
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
      parameters:
      - name: id
        in: path
        required: true
        schema: *id001
        description: Identificador id
        example: 11111111-1111-4111-8111-111111111111
    patch:
      operationId: api042
      x-teralya-api-code: API-042
      x-teralya-cu:
      - CU-027
      x-teralya-pt:
      - PT-ADM-008
      - PT-ADM-009
      tags:
      - Administración
      summary: Actualizar incidencia
      description: Contrato normativo de API-042.
      security:
      - BearerOpaque: []
      responses:
        '200':
          description: Operación completada
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentDetail'
              examples:
                success:
                  value:
                    id: 11111111-1111-4111-8111-111111111111
                    tipo: aaaaa
                    estado: abierta
                    fecha: '2026-07-13T12:00:00Z'
                    recurso_relacionado:
                      tipo: pedido
                      id: 11111111-1111-4111-8111-111111111111
                    descripcion: aaaaa
                    updated_at: '2026-07-13T12:00:00Z'
        '400':
          description: Validation Error
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/validation-error
                    title: Validation Error
                    status: 400
                    detail: La solicitud no pudo completarse.
                    code: VALIDATION_ERROR
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '401':
          description: Authentication Required
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            WWW-Authenticate:
              description: Desafío Bearer según RFC 6750
              schema:
                type: string
                const: Bearer
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/authentication-required
                    title: Authentication Required
                    status: 401
                    detail: La solicitud no pudo completarse.
                    code: AUTHENTICATION_REQUIRED
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '403':
          description: Forbidden
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/forbidden
                    title: Forbidden
                    status: 403
                    detail: La solicitud no pudo completarse.
                    code: FORBIDDEN
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '404':
          description: Resource Not Found
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/resource-not-found
                    title: Resource Not Found
                    status: 404
                    detail: La solicitud no pudo completarse.
                    code: RESOURCE_NOT_FOUND
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '409':
          description: Conflict
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/conflict
                    title: Conflict
                    status: 409
                    detail: La solicitud no pudo completarse.
                    code: CONFLICT
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '500':
          description: Internal Error
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/internal-error
                    title: Internal Error
                    status: 500
                    detail: La solicitud no pudo completarse.
                    code: INTERNAL_ERROR
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
      parameters:
      - name: id
        in: path
        required: true
        schema: *id001
        description: Identificador id
        example: 11111111-1111-4111-8111-111111111111
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidentStatePatch'
            examples:
              success:
                value:
                  estado_destino: abierta
...
components:
  securitySchemes:
    BearerOpaque:
      type: http
      scheme: bearer
      description: Token opaco de 256 bits; TTL absoluto 8 horas.
  headers:
    XRequestId:
      description: Identificador UUID de trazabilidad
      schema: *id001
  schemas:
    FieldError:
      type: object
      additionalProperties: false
      properties:
        field: &id013
          type: string
          minLength: 1
          maxLength: 200
        code: *id006
        message: &id007
          type: string
          minLength: 1
          maxLength: 500
      required:
      - field
      - code
      - message
    Problem:
      type: object
      additionalProperties: false
      properties:
        type: &id016
          type: string
          format: uri
          minLength: 1
          maxLength: 2048
        title: *id002
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail: *id007
        instance:
          type: string
          minLength: 1
          maxLength: 2048
        code: *id006
        request_id: *id001
        field_errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
          maxItems: 50
        retryable:
          type: boolean
      required:
      - type
      - title
      - status
      - detail
      - code
      - request_id
      - retryable
...
    IncidentStatePatch:
      type: object
      additionalProperties: false
      properties:
        estado_destino: *id032
      required:
      - estado_destino
    RelatedResource:
      type: object
      additionalProperties: false
      properties:
        tipo:
          type: string
          enum:
          - pedido
          - subpedido
          - bodega
          - vino
        id: *id001
      required:
      - tipo
      - id
    IncidentSummary:
      type: object
      additionalProperties: false
      properties:
        id: *id001
        tipo: *id009
        estado: *id032
        fecha: *id018
        recurso_relacionado:
          $ref: '#/components/schemas/RelatedResource'
      required:
      - id
      - tipo
      - estado
      - fecha
      - recurso_relacionado
    IncidentDetail:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        tipo:
          type: string
          minLength: 1
          maxLength: 100
        estado:
          type: string
          enum:
          - abierta
          - en_revision
          - resuelta
          - cerrada
        fecha:
          type: string
          format: date-time
        recurso_relacionado:
          $ref: '#/components/schemas/RelatedResource'
        descripcion: *id015
        updated_at: *id018
      required:
      - id
      - tipo
      - estado
      - fecha
      - recurso_relacionado
      - descripcion
      - updated_at
...
    PageIncidentSummary:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/IncidentSummary'
          maxItems: 100
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          minimum: 1
          maximum: 100
        total_items:
          type: integer
          minimum: 0
        total_pages:
          type: integer
          minimum: 0
      required:
      - items
      - page
      - page_size
      - total_items
      - total_pages
...